+++ src/frontend/src/views/index.ts
@@ -7,3 +7,5 @@ export * from "./AgentCanvasView";
 export * from "./AgentMarketplaceView";
 export * from "./WeatherView";
 export * from "./LandingPage";
+export * from "./LoginView";
+export * from "./RegisterView";